# api/docs/swagger.yaml
basePath: /
definitions:
...
  api.ExecChallengeRequest:
    properties:
      command:
        items:
          type: string
        type: array
      instance_id:
        type: string
      timeout_seconds:
        type: integer
    type: object
  api.ExecChallengeResponse:
    properties:
      exit_code:
        type: integer
      stderr:
        type: string
      stdout:
        type: string
      truncated:
        type: boolean
    type: object
...
  api.HTTPErrorResp:
    properties:
      error:
        example: Error occured while veifying the challenge.
        type: string
    type: object
  api.HTTPPlainResp:
    properties:
      message:
        example: Messsage in response to your request
        type: string
    type: object
...
info:
  contact:
    email: contact@sdslabs.co.in
    name: SDSLabs
    url: https://chat.sdslabs.co
  description: Authenticated API for Beast CTF challenge deployment and competition
    services.
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  title: Beast API
  version: "0.2"
paths:
...
  /api/config/challenge-info:
    post:
      consumes:
      - application/json
      description: Updates challenge info in the database, located at ~/.beast/beast.db.
      parameters:
      - description: Challenge Name
        in: formData
        name: name
        required: true
        type: string
      - description: Challenge's description
        in: formData
        name: desc
        type: string
      - description: Challenge's points
        in: formData
        name: points
        type: string
      - description: Challenge's flag
        in: formData
        name: flag
        type: string
      - description: Challenge's tags
        in: formData
        name: tags
        type: string
      - description: Challenge's ports
        in: formData
        name: ports
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.HTTPPlainResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.HTTPPlainResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.HTTPErrorResp'
      summary: Updates challenge info in the database, located at ~/.beast/beast.db.
      tags:
      - config
...
  /api/manage/challenge/{name}/exec:
    post:
      consumes:
      - application/json
      parameters:
      - description: Challenge name
        in: path
        name: name
        required: true
        type: string
      - description: Bounded exec request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.ExecChallengeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ExecChallengeResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.HTTPErrorResp'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.HTTPErrorResp'
      security:
      - ApiKeyAuth: []
      summary: Execute an argument-vector command in an owned challenge container
      tags:
      - manage
...
schemes:
- https
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"