# api/response.go
package api

import (
	"time"
)

type HTTPPlainResp struct {
	Message string `json:"message" example:"Messsage in response to your request"`
}

type HTTPPlainMapResp struct {
	Messages map[string]string `json:"messages" example:"{\"name1\": \"message1\", \"name2\": \"message2\"}"`
}

type HTTPErrorResp struct {
	Error string `json:"error" example:"Error occured while veifying the challenge."`
}

type HTTPAuthorizeResp struct {
	Token   string `json:"token" example:"YOUR_AUTHENTICATION_TOKEN"`
	Role    string `json:"role" example:"author"`
	Message string `json:"message" example:"Response message"`
}

type AvailableImagesResp struct {
	Message string   `json:"message" example:"Available Base images."`
	Images  []string `json:"images" example:"ubuntu:24.04,debian:bookworm"`
}

type PortsInUseResp struct {
	MinPortValue uint32   `json:"port_min_value" example:"10000"`
	MaxPortValue uint32   `json:"port_max_value" example:"20000"`
	PortsInUse   []uint32 `json:"ports_in_use" example:"10001,10003,10010"`
}

type ChallengeStatusResp struct {
	Name      string    `json:"name" example:"Web Challenge"`
	Status    string    `json:"status" example:"deployed"`
	UpdatedAt time.Time `json:"updated_at" example:"2018-12-31T22:20:08.948096189+05:30"`
}

type LogsInfoResp struct {
	Stdout string `json:"stdout" example:"[INFO] Challenge is starting to deploy"`
	Stderr string `json:"stderr" example:"[ERROR] Challenge deployment failed."`
}

type ChallengeDescriptionResp struct {
	Name   string `json:"name" example:"Web Challenge"`
	Author string `json:"author" example:"Fristonio"`
	Desc   string `json:"desc" example:"Challenge Description"`
}

type NotificationResp struct {
	ID        uint      `json:"id" example:"3"`
	Title     string    `json:"title" example:"CTF is live now!"`
	Desc      string    `json:"desc" example:"Notification Description"`
	UpdatedAt time.Time `json:"updated_at" example:"2018-12-31T22:20:08.948096189+05:30"`
}

type UserResp struct {
	Id         uint                 `json:"id" example:"5"`
	Username   string               `json:"username" example:"CTF is live now!"`
	Role       string               `json:"role" example:"author"`
	Status     uint                 `json:"status" example:"0"`
	Score      uint                 `json:"score" example:"750"`
	Rank       int64                `json:"rank" example:"15"`
	Email      string               `json:"email" example:"fristonio@gmail.com"`
	Challenges []ChallengeSolveResp `json:"challenges"`
}

type UsersResp struct {
	Id       uint   `json:"id" example:"5"`
	Username string `json:"username" example:"CTF is live now!"`
	Role     string `json:"role" example:"author"`
	Status   uint   `json:"status" example:"0"`
	Score    uint   `json:"score" example:"750"`
	Email    string `json:"email" example:"fristonio@gmail.com"`
	Rank     int64  `json:"rank" example:"15"`
}

type ChallengeSolveResp struct {
	Id       uint      `json:"id" example:"4"`
	Name     string    `json:"name" example:"Web Challenge"`
	Category string    `json:"category" example:"bare"`
	Tags     []string  `json:"tags" example:"pwn,misc"`
	SolvedAt time.Time `json:"solvedAt"`
	Points   uint      `json:"points" example:"50"`
}

type UserSolveResp struct {
	Id       uint      `json:"id" example:"5"`
	Username string    `json:"username" example:"fristonio"`
	SolvedAt time.Time `json:"solvedAt"`
	Flag     string    `json:"flag" example:"flag{example_flag}"`
	Correct  bool      `json:"correct" example:"true"`
}

type HintInfo struct {
	ID     uint `json:"id"`
	Points uint `json:"points"`
}

type HintResponse struct {
	Description string `json:"description" example:"This is a hint"`
	Points      uint   `json:"points" example:"10"`
}

type ChallengeMetadata struct {
	ChallId            uint      `json:"id" example:"0"`
	Name               string    `json:"name" example:"Web Challenge"`
	Tags               []string  `json:"tags" example:"pwn,misc"`
	Points             uint      `json:"points" example:"50"`
	Difficulty         string    `json:"difficulty" example:"easy"`
	SolvesNumber       uint16    `json:"solvesNumber" example:"100"`
	SolveStatus        bool      `json:"solveStatus" example:"True"`
	CreatedAt          time.Time `json:"createdAt"`
	DeployedStatus     string    `json:"deployedStatus" example:"deployed"`
	PreRequisite       []string  `json:"preRequisite" example:"chall1,chall2"`
	Instanced          bool      `json:"instanced" example:"false"`
	InstanceExpiration int64     `json:"instanceExpiration" example:"300"`
}

type Challenge struct {
	ChallengeMetadata

	Description     string     `json:"description" example:"A simple web challenge"`
	Hints           []HintInfo `json:"hints"`
	Category        string     `json:"category" example:"web"`
	Assets          []string   `json:"assets" example:"image1.png,zippy.zip"`
	AdditionalLinks []string   `json:"additionalLinks" example:"https://link1.example,https://link2.example"`
	PreviousTries   int        `json:"previousTries" example:"3"`
	MaxAttemptLimit int        `json:"maxAttemptLimit" example:"5"`
	DeployedLink    string     `json:"deployedLink" example:"beast.sdslabs.co or ip:port"`
}

// ChallengeDetails is an extended struct for admin use, containing additional fields.
type AdminChallenge struct {
	Challenge
	DynamicFlag bool   `json:"dynamicFlag" example:"true"`
	Flag        string `json:"flag"`
}

type ChallengePreviewResp struct {
	Name            string   `json:"name" example:"Web Challenge"`
	Category        string   `json:"category" example:"web"`
	Tags            []string `json:"tags" example:"pwn,misc"`
	Assets          []string `json:"assets" example:"image1.png,zippy.zip"`
	AdditionalLinks []string `json:"additionalLinks" example:"https://link1.example,https://link2.example"`
	MaxAttemptLimit int      `json:"maxAttemptLimit" example:"5"`
	PreReqs         []string `json:"preRequisite" example:"web-php,simple"`
	Ports           []uint32 `json:"ports" example:"3001,3002"`
	Desc            string   `json:"description" example:"A simple web challenge"`
	Points          uint     `json:"points" example:"50"`
	DeployedLink    string   `json:"deployedLink" example:"beast.sdslabs.co"`
}

type SubmissionResp struct {
	UserId    uint      `json:"user_id" example:"3"`
	Username  string    `json:"username" example:"fristonio"`
	ChallId   uint      `json:"chall_id" example:"3"`
	ChallName string    `json:"name" example:"Web Challenge"`
	Flag      string    `json:"flag" example:"flag{@#$}"`
	SolvedAt  time.Time `json:"solvedAt"`
	Success   bool      `json:"success" example:"true"`
	Cheating  bool      `json:"cheating" example:"false"`
}

type FlagSubmitResp struct {
	Message string `json:"message" example:"Your answer is correct"`
	Success bool   `json:"success" example:"true"`
}

type UsersStatisticsResp struct {
	TotalRegisteredUsers uint `json:"total_registered_users" example:"120"`
	BannedUsers          uint `json:"banned_users" example:"60"`
	UnbannedUsers        uint `json:"unbanned_users" example:"60"`
}

type CompetitionInfoResp struct {
	Name         string `json:"name" example:"fristonio"`
	About        string `json:"about" example:"This is a CTF competition"`
	Prizes       string `json:"prizes" example:"1st and 2nd place winners will get $10K"`
	StartingTime string `json:"starting_time"`
	EndingTime   string `json:"ending_time"`
	TimeZone     string `json:"timezone" example:"Asia/Calcutta: UTC +05:30"`
	LogoURL      string `json:"logo_url"`
}

type TagInfoResp struct {
	Tags []string `json:"tags"`
}

// UserCountResp represents the response for contestant user count
type UserCountResp struct {
	UserCount int64 `json:"user_count"`
}